==> 50_chatbot.ipynb <==
import importlib
import src.chat_bot
importlib.reload(src.chat_bot)
from src.chat_bot import ExconManual

ad = True
#ad = False # = ADLA chat
if ad:
    print("Loaded Authorised Dealer Manual")
else:
    print("Loaded ADLA manual")

logging_level = 15
#logging_level = 20
log_file = ''

if ad:
    path_to_manual_as_csv_file = "./inputs/ad_manual.csv"
    path_to_definitions_as_parquet_file = "./inputs/ad_definitions.parquet"
    path_to_index_as_parquet_file = "./inputs/ad_index.parquet"
    chat_for_ad = True
else:
    path_to_manual_as_csv_file = "./inputs/adla_manual.csv"
    path_to_definitions_as_parquet_file = "./inputs/adla_definitions.parquet"
    path_to_index_as_parquet_file = "./inputs/adla_index.parquet"
    chat_for_ad = False


excon = ExconManual(path_to_manual_as_csv_file, path_to_definitions_as_parquet_file, path_to_index_as_parquet_file, chat_for_ad = chat_for_ad, log_file=log_file, logging_level=logging_level)


print(excon.get_regulation_detail("A.3 (B)(ii)(a)"))

#model_to_use="gpt-3.5-turbo"
#model_to_use = "gpt-4-1106-preview"
model_to_use="gpt-4"
#model_to_use="gpt-3.5-turbo-16k"

import pandas as pd
df = pd.read_parquet(path_to_index_as_parquet_file, engine="pyarrow")


tmp_df = df[(df["section"].str.startswith("B.1(D)(i)")) & (df["source"] == "question")]
#tmp_df = df[(df["section"].str.startswith("B.1(D)(i)")) & (df["source"] == "heading")]
tmp_df
# for index, row in tmp_df.iterrows():
#     print(f'{index}: {row["text"]}\n')


new_question = index_2
new_embedding = index_2_e
source = "manual"

#index_to_add = len(df)
df.loc[index_to_add,'text'] = new_question
df.loc[index_to_add,'Embedding'] = new_embedding
df.loc[index_to_add,'source'] = source
df.loc[index_to_add,'section'] = 'B.1(D)(i)'


#df.loc[index_to_add,'source']

df.loc[index_to_add,'text']
#df.loc[index_to_add,'Embedding']

df

df.to_parquet(path_to_index_as_parquet_file, engine="pyarrow")

heading_1 = "Payment for imports. Terms of payment for imports. Advance payments"
heading_1_e = get_ada_embedding(heading_1)

p_1 = "Advance Payments for Imports"
p_1_e = get_ada_embedding(p_1)
p_2 = "Details on Advance Payment Conditions under Import Payment Terms"
p_2_e = get_ada_embedding(p_2)
p_3 = "Exploring the Mechanism of Advance Payments in Import Trade"
p_3_e = get_ada_embedding(p_3)


from src.embeddings import get_ada_embedding
from scipy.spatial import distance

index_1 = "What are the guidelines for making advance payments for imports in foreign currency?"
index_1_e = get_ada_embedding(index_1)
index_2 = "What are the guidelines for making advance payments?"
index_2_e = get_ada_embedding(index_2)
index_3 = "What are the guidelines for making advance payments for imports?"
index_3_e = get_ada_embedding(index_3)

#question = "What documentation is required for an advance payments?"
question = "What documentation is required for an advance payments for imports?"
question_e = get_ada_embedding(question)




print(f'question to index_1  : {distance.cosine(question_e, index_1_e):.4f}')
print(f'question to index_2  : {distance.cosine(question_e, index_2_e):.4f}')
print(f'question to index_3  : {distance.cosine(question_e, index_3_e):.4f}')
print(f'question to heading_2: {distance.cosine(question_e, heading_1_e):.4f}')
print(f'question to p_1      : {distance.cosine(question_e, p_1_e):.4f}')
print(f'question to p_2      : {distance.cosine(question_e, p_2_e):.4f}')
print(f'question to p_3      : {distance.cosine(question_e, p_3_e):.4f}')


#user_context = "A non-resident wants to buy residential property in South Africa. What are the restrictions around doing this?"
#user_context = "A South African citizen is emigrating. What is the process they need to follow to take all their cash offshore?"
#user_context = "What is the Foreign investment limits for South African institutional investors?"
#user_context = "Can a individual transfer his shares in a dual listed company to his offshore portfolio?"
#user_context = "Can cash balances for individuals who are no longer tax resident be transferred offshore?"
#user_context = "How do I need to report outbound cross-boarder payments?"
#user_context = "Are there different categories for ALDAs?"
#user_context = "Who can trade gold?"

#user_context = "What documentary evidence is required?"
#user_context = "What does cemad say about travel services?"
#user_context = "Can I import gold?"
#user_context = "how much money can an individual take offshore?"
#user_context = "How much can an individual take offshore in a year?"
#user_context = "what is the limit for commodities"
#user_context = "what does the regulation state about travel?"
#user_context = "How much can adult residents use for travel?"

#user_context = "If I want to offer cross-border payments, what license do I need?"

#user_context = "What documentation is required for an advance payments?"
user_context = "What documentation is required for an advance payments for imports?"

excon.reset_conversation_history()
excon.user_provides_input(user_context, threshold=0.15, model_to_use=model_to_use, temperature=0, max_tokens=300)
print(excon.messages[-1]["content"])


excon.messages

user_context = "hi"
#user_context = "but what category do I need to be?"
excon.user_provides_input(user_context, threshold=0.15, model_to_use=model_to_use, temperature=0, max_tokens=300)
print(excon.messages[-1]["content"])

import importlib
import src.valid_index
importlib.reload(src.valid_index)
from src.valid_index import get_excon_manual_index
import re

index = get_excon_manual_index()


s = " B.4(B)(ii)(a) of the Manual states that foreign currency may be made available within the single discretionary allowance limit of R1 million per calendar year to eligible parties. However, the specific limits for adults are not mentioned in the provided sections. Please refer to the relevant section(s) for more information on the limits for travel allowances for adults."
#s = "B.4(B)(ii)(a)"

print(index.extract_valid_reference(s))


patterns = [
        r'[A-Z]\.\d{0,2}',
        r'\([A-Z]\)',
        r'\(\b(i|ii|iii|iv|v|vi|vii|viii|ix|x|xi|xii|xiii|xiv|xv|xvi|xvii|xviii|xix|xx|xxi|xxii|xxiii|xxiv|xxv|xxvi|xxvii)\b\)',
        r'\([a-z]\)',
        r'\([a-z]{2}\)',
        r'\((?:[1-9]|[1-9][0-9])\)'
    ]
matches = [re.findall(pattern, s) for pattern in patterns]
matches

match = re.search(r'\([a-z]\)', s)
match

s[match.end()]

excon.get_regulation_detail("B.5(B)(iii)")

print(excon._create_system_message())

import pandas as pd
df_text_all = pd.read_parquet(path_to_index_as_parquet_file, engine='pyarrow')
df_text_all

excon.get_regulation_detail('I.(1)(E)(ii)')

#user_context = "What is a CFC Account?"
#user_context = "I am submitting documentation to the financial surveillance department. What information do I need to include in my submission?" # (A.3 (B))
#user_context = "Can a bank provide foreign currency denominated facilities to a South African Corporate?" #  (A.3 (B))
#user_context = "For how long does the approval for a facility remain valid?" #  (A.3 (B))
#user_context = "Other than authorised transaction requests, are there requirements for authorised dealers to submit information to the financial surveillance department?" #  (A.3 (B))
#user_context = "I have been asked to act on an approval issued to another bank. Can I do that?" #  (A.3 (B))
#####user_context = "--- Same as above but use "Authorised Dealer" instead of "Bank"
#user_context = "If I am concerned that as client may in breach of regulations, what should I do?" #  (A.3 (B))
#user_context = "What constitutes 'documentary evidence' for a foreign currency transaction?" #  (A.3 (C))
#user_context = "Is Swaziland part of the CMA?" #  (A.3 (E))
#user_context = "A client wants to buy Namibian Dollars. What approvals do they need before they can do this?" #  (A.3 (E))
#user_context = "A person from Lesotho wants to buy dollars from me. Can I sell them the dollars?" #  (A.3 (E))
#user_context = "A person from Namibia wants to buy rand from me. Can I sell them the rand?" #  (A.3 (E))
#user_context = "A treasury management company wants to trade dollars with me. What do I need to check before I can do this?" #  (A.4 (B))
#user_context = "A client wants to import goods and pay in foreign currency. What documentation do I need before I can assist with the transaction" #  (B.1 (B))
#user_context = "Over and above the price of the goods, what additional costs can be covered in foreign currency?" #  (B.1 (C))
#user_context = "Can advance payment be made in a foreign currency to cover imported goods?" #  (B.1 (D)(i))

#user_context = "Can advance payment be made in a foreign currency to cover imported capital goods?" #  (B.1 (D)(ii))
#user_context = "When paying for imports in cash, what documentation is required before providing the foreign currency?" #  (B.1 (D)(iii))

#user_context = "Can imported goods be financed over more than 12 months?" #  (B.1 (D)(iv))
#user_context = "How much money can an individual invest internationally?" #  ?
#user_context = "How much money can an individual take offshore?" #  ?

#user_context = "A media company want to list offshore. Are there specific conditions for this?"
# user_context = "How many rands can a resident take offshore in cash?"
#user_context = "Can I grant a foreign company a Rand facility to pay for goods they wish to purchase from a South African exporter?"
#user_context = "Can I support a non-residential entity with short-term rand trade finance for their exports from South Africa?"
#user_context = "Hi"
#user_context = "Who can trade gold?"
#user_context = "are there specific exchange control rules for religious organizations?"
#user_context = "How much foreign currency can a charity take offshore each year?"
